+++ packages/server/api/models/index.ts
@@ -1,17 +1,7 @@
 import { Users } from './users';
-import { Forums } from './forums';
-import { Topics } from './topics';
-import { Messages } from './messages';
-import { Reactions } from './reactions';
-import { MessagesReactions } from './messagesReactions';
-import { Themes } from './themes';
+import type { TUser } from './users';
 
 export {
   Users,
-  Forums,
-  Topics,
-  Messages,
-  Reactions,
-  MessagesReactions,
-  Themes,
+  TUser,
 };
