+++ packages/server/authMiddleware/authorization.ts
@@ -1,10 +1,8 @@
 import { createProxyMiddleware } from 'http-proxy-middleware';
 import type { RequestHandler } from 'http-proxy-middleware';
-import { yandexEndpoint } from './constants';
-import { filterCookies } from './filterCookies';
-import { yandexProxyResponseHandler } from 'server/api/utils';
+import { yandexEndpoint } from 'server/authMiddleware/constants';
+import { filterCookies, yandexProxyResponseHandler } from 'server/authMiddleware/utils';
 
-// Прокси для запросов к Яндекс Апи
 export const yandexProxyAll = (): RequestHandler => {
   return createProxyMiddleware({
     changeOrigin: true,
@@ -18,8 +16,6 @@ export const yandexProxyAll = (): RequestHandler => {
     },
   });
 };
-
-// Прокси для запросов к Яндекс Апи
 export const yandexProxyUserInfoOnly = (): RequestHandler => {
   return createProxyMiddleware({
     changeOrigin: true,
