+++ packages/server/authMiddleware/index.ts
@@ -1,0 +1,8 @@
+import { yandexProxyAll, yandexProxyUserInfoOnly } from 'server/authMiddleware/authorization';
+import { yandexCheckAuthorization } from 'server/authMiddleware/checking';
+
+export {
+  yandexProxyAll,
+  yandexProxyUserInfoOnly,
+  yandexCheckAuthorization,
+};
