+++ packages/server/authMiddleware/utils/filterCookies.ts
