+++ packages/server/server.ts
@@ -1,12 +1,17 @@
-import { yandexProxyAll, yandexProxyUserInfoOnly, yandexCheckAuthorization } from 'server/api/auth';
+import {
+  yandexProxyAll,
+  yandexProxyUserInfoOnly,
+  yandexCheckAuthorization,
+} from 'server/authMiddleware';
 import type { ViteDevServer } from 'vite';
 import { createServer as createViteServer } from 'vite';
 import cors from 'cors';
 import express from 'express';
 import path from 'path';
-import { forumApi } from 'server/api/forum';
+import { forumApiHandler } from 'server/api/forum';
 import { getSsrPath, ssrContent } from './ssr';
 import cookieParser from 'cookie-parser';
+
 export async function startServer(isDev: boolean, port: number) {
   const app = express();
   app.use(cors());
@@ -22,22 +27,21 @@ export async function startServer(isDev: boolean, port: number) {
     const distPath = path.dirname(require.resolve('client/dist/index.html'));
     app.use('/assets', express.static(path.resolve(distPath, 'assets')));
   }
-
   app.use('/api/v2/auth/user', yandexProxyUserInfoOnly());
   app.use('/api/v2', yandexProxyAll());
-
   app.use('/api/forum', async (req, res) => {
     try {
       const authUserData = await yandexCheckAuthorization(req);
-      if (!authUserData.isAuth) {
+      if (!authUserData.isAuth || !authUserData.user) {
         res.sendStatus(403);
         return;
       }
       app.use(express.json());
-      await forumApi(req, res);
+      await forumApiHandler(req, res, authUserData.user);
     } catch (e) {
-      console.log(e);
-      res.sendStatus(500);
+      if (!res.headersSent) {
+        res.sendStatus(500);
+      }
     }
   });
   app.use('*', cookieParser() as any, async (req, res, next) => {
